Add at() and [] operators to the hashmaps
They return pointers instead of references to avoid the standard api where accessing creates an object, and not finding the object throws an exception
These functions return nullptr when not found, a much easier and nicer way to use them

+++ include/crstl/platform/process_posix.h
@@ -164,7 +164,7 @@ crstl_module_export namespace crstl
 
 				if (bytes_read >= 0)
 				{
-					return process_size(bytes_read);
+					return process_size((size_t)bytes_read);
 				}
 			}
